sync the pick up
create new kitchenobjectSoList scriptableobject to get index of kitchenobject to pass index in serverRpc method
SpawnKitchenObjectServerRpc

the index will be used get kitchenobject . it is need to pass barrer of serverRpc which doesn't take kitchenObjectSo as parameter.

In similarly way KitchObjectParent is not taken. therefore NetworkObjectReference kitchenObjectParentNetworkObjectReference is used to get networkobject from TryGet method and used networkobject to getcomponent IKitchenObjectParent