sync kitchenobject parent
to sync the kitchen object parent,
create the new script called the FollowTransform which change transform of kitchenobject according to input transfrom.

input of FollowTransform is KitchenObjectParent.GetKitchenObjectFollowTransform() that mean empty game object in player and countertopPoint in counter

Reason for using this Follow Transform because network object can't be parent to non-Network object and network object.

Thre is was problem while pickup the pateKitchenObject because Awake method in kitchenobject and Awake method on plateKicthen object were callled same time.

There Awake method in kitchen object is change in  protected virtual and protected override Awake in plateKitchenobject  and called  awake method of kitchenobject from awake method of plateKitchenObject using the  base.Awake();