sync Cutting Counter
sync the interact with ServerRpc and ClientRpc but there are error or null error in  player.GetKitchenObject() because client take time to play therefore  player.GetKitchenObject() was store in kitchenObject.
Interact method was divided in serverRpc and ClientRpc to reset the cutting progress after cutting is over and replace the object.

serverRpc and Client Rpc in Interact is used after the SetKitchenObjectParent beause it already using ServerRpc and ClientRpc
Cutting InteractAlternate using ServerRpc and ClientRpc  to track the  cutting progress of kitchenObject and boardcast all client