add player visual and ready UI in Character Select scene
add player visual when player join and show Ready UI when player is ready.

to shown when player is join with visual player with help playerindex. playerindex is used to check if player index in exist in KitchenGameMultiplayer playerDataNetworkList or not

playerDataNetworkList add using NetworkManager OnClientConnectedCallback methid  or event in PlayerData struct

to check player is ready for not through OnReadyChanged event. this event is inoke when playeid  is true in playerReadyDictionary and SetPlayerReadyClientRpc to client and invoke it and used IsPlayerReady to  return bool to CharacterSelectPlayer and used OnReadyChanged  event to check playerdata  or clientid  from KitchenGameMultiplayer  is true or not